WWE-772 [Agent Api Simulator] Issues with multi-tabs
Reason:
The issue was caused because of the algorithm which works when the user creates new tab. When user run new tab, the program creates new session and switch it with the cometd session already existed and the server-side events were published to the last opened tab.
Solution:
Instead of single session for last opened tab API simulator will use an array with the all opened tabs sessions. When the user creates new tab, sessions of the old tabs will be saved and will continue to recieve published events.