[Agent-Api-Simulator] The Interaction Bar shows wrong party member after agent transfers an internal call
The reason of the problem was in incorrect combinations of usernames and releated calls.
It also was no changing of destination user during the transfer

Now it is correctly publish call events with correct combinations of usernames and calls.
It is also add number of agent who gets the transfer to participants[0] where saved call destination